[DS] REFACTOR | edit file and class names
